tests: add e2e + system test for backing types
